First pass at the User Guide and imported into Nmap
